-Fixed Combat Dagger Attack Damager Attribute -Fixed Combat Dagger (Or Combat Form Weapons) not showing particles when held -Added A few other mixins to in relation to the above problems -Prob many more things I cant remember

+++ src/main/java/net/kenji/kenjiscombatforms/item/custom/fist_forms/EnderFormItem.java
@@ -35,7 +35,7 @@ public class EnderFormItem extends BaseFinalFormClass {
             ImmutableMultimap.Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
             builder.putAll(super.getDefaultAttributeModifiers(slot));
 
-            int baseDamage = EpicFightCombatFormsCommon.BASIC_FORM_BASE_DAMAGE.get();
+            double baseDamage = EpicFightCombatFormsCommon.BASIC_FORM_BASE_DAMAGE.get();
             double damageMultiplier = EpicFightCombatFormsCommon.LEVEL1_DAMAGE_MULTIPLIER.get();
             double finalDamage = baseDamage * damageMultiplier; // Subtracting 2 because Minecraft adds it
