+++ src/main/java/net/kenji/kenjiscombatforms/item/custom/fist_forms/power_form/PowerFist3Item.java
@@ -46,7 +46,7 @@ public class PowerFist3Item extends BasePowerClass {
         if (slot == EquipmentSlot.MAINHAND) {
             ImmutableMultimap.Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
 
-            int baseDamage = EpicFightCombatFormsCommon.POWER_FORM_BASE_DAMAGE.get();
+            double baseDamage = EpicFightCombatFormsCommon.POWER_FORM_BASE_DAMAGE.get();
             double damageMultiplier = EpicFightCombatFormsCommon.LEVEL3_DAMAGE_MULTIPLIER.get();
             double finalDamage = baseDamage * damageMultiplier; // Subtracting 2 because Minecraft adds it
