+++ src/main/java/net/kenji/kenjiscombatforms/item/custom/fist_forms/swift_form/SwiftFist2Item.java
@@ -42,7 +42,7 @@ public class SwiftFist2Item extends BaseSwiftClass {
         if (slot == EquipmentSlot.MAINHAND) {
             ImmutableMultimap.Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
 
-            int baseDamage = EpicFightCombatFormsCommon.SWIFT_FORM_BASE_DAMAGE.get();
+            double baseDamage = EpicFightCombatFormsCommon.SWIFT_FORM_BASE_DAMAGE.get();
             double damageMultiplier = EpicFightCombatFormsCommon.LEVEL2_DAMAGE_MULTIPLIER.get();
             double finalDamage = baseDamage * damageMultiplier; // Subtracting 2 because Minecraft adds it
