+++ src/main/java/net/kenji/kenjiscombatforms/item/custom/fist_forms/swift_form/SwiftFistItem.java
@@ -43,7 +43,7 @@ public class SwiftFistItem extends BaseSwiftClass {
         if (slot == EquipmentSlot.MAINHAND) {
             ImmutableMultimap.Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
 
-            int baseDamage = EpicFightCombatFormsCommon.SWIFT_FORM_BASE_DAMAGE.get();
+            double baseDamage = EpicFightCombatFormsCommon.SWIFT_FORM_BASE_DAMAGE.get();
             double damageMultiplier = EpicFightCombatFormsCommon.LEVEL1_DAMAGE_MULTIPLIER.get();
             double finalDamage = baseDamage * damageMultiplier; // Subtracting 2 because Minecraft adds it
